requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-community
sentence-transformers
qdrant-client
flashrank
faiss-cpu